feat(grammar): add support for floating point literals
REPORTING-0118

+++ src/aggregator/interpreter.js
@@ -51,7 +51,7 @@ module.exports = (parser) => {
     }
 
     numberExpression(ctx) {
-      return parseInt(ctx.numberLiteral[0].image);
+      return parseFloat(ctx.numberLiteral[0].image);
     }
 
     stringExpression(ctx) {
